+++ lib/chainsync/chainsync/db/api/__init__.py
@@ -1,3 +1,3 @@
 """Api server for the chainsync database."""
 from .api_interface import balance_of, register_username
-from .flask_server import launch_flask
+from .api_server import launch_flask
